requirements.txt
opencv-python>=4.8.0
gspread>=5.12.0
google-auth>=2.23.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
imutils>=0.5.4
